+++ src/Grace.Shared/Validation/Repository.Validation.fs
@@ -8,12 +8,14 @@ open System
 
 module Repository =
 
+    /// Checks that the visibility value provided exists in the RepositoryVisibility type.
     let visibilityIsValid (visibility: string) (error: RepositoryError) =
         match Utilities.discriminatedUnionFromString<RepositoryVisibility>(visibility) with
         | Some visibility -> Ok () |> returnTask
         | None -> Error error |> returnTask
 
-    let daysIsValid (days: float) (error: RepositoryError) =
+    /// Checks that the number of days is between 0.0 and 65536.0.
+    let daysIsValid (days: double) (error: RepositoryError) =
         if days < 0.0 || days >= 65536.0 then
             Error error |> returnTask
         else
